Cherry-pick to stabilization: Enable testing using SDK (#18883)
* Installer now supports tests and a new template adds test to existing projects (#18744)

* Adds the ability to make an installer which supports user tests

With these changes, the installer can be built with -DLY_DISABLE_TEST_MODULES=ON
but still generate an installer that users can use to run their OWN tests.

You can also build your own project based on the source code (not just the
installer) and use -DLY_DISABLE_TEST_MODULES=ON to disable the tests in the
engine, but leave your own tests enabled in your project.

This works on installer as well as source builds - LY_DISABLE_TEST_MODULES
disables tests in the engine.  They will only disable gem or project tests
if those gems or projects uses `if (Y_DISABLE_TEST_MODULES)` in their
own CMakeLists.txt, which they do not have to do, they can use their
own custom variable for example, to leave their tests on, but engine
tests off.

Tested on linux with installer and source builds, with a project
that has a googletest as well as a pytest (both work).

Both work with LY_DISABLE_TEST_MODULES=ON and OFF.
note that with LY_DISABLE_TEST_MODULES=ON, the engine tests are added
to your tests, if you're using a source build.  But with a installer build
the tests are not added to your tests since the installer is built with
them off and no project added.

* Adds a template for adding tests to any project
* Adds license snippet to tests
* Fixes android build

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Disable google test support for monolythic (#18850)

Signed-off-by: Mateusz Wasilewski <mateusz.wasilewski@robotec.ai>

* Enable common test utility targets for sdk (#18862)

Signed-off-by: Mateusz Wasilewski <mateusz.wasilewski@robotec.ai>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Signed-off-by: Mateusz Wasilewski <mateusz.wasilewski@robotec.ai>
Co-authored-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Co-authored-by: Mateusz Wasilewski <mateusz.wasilewski@robotec.ai>